+++ click_extra/commands.py
@@ -49,6 +49,7 @@ from .context import ExtraContext
 from .envvar import clean_envvar_id, param_envvar_ids
 from .execution import TimerOption
 from .logging import VerboseOption, VerbosityOption
+from .man_page import ManOption
 from .parameters import ExtraOption, ShowParamsOption
 from .table import TableFormatOption
 from .theme import ThemeOption
@@ -88,6 +89,7 @@ def default_extra_params() -> list[click.Option]:
     #. ``--table-format FORMAT``
     #. ``--verbosity LEVEL``
     #. ``-v``, ``--verbose``
+    #. ``--man``
     #. ``--version``
     #. ``-h``, ``--help``
         .. attention::
@@ -128,6 +130,7 @@ def default_extra_params() -> list[click.Option]:
         TableFormatOption(),
         VerbosityOption(),
         VerboseOption(),
+        ManOption(),
         ExtraVersionOption(),
         # @click.decorators.help_option(),
     ]