+++ click_extra/pytest.py
@@ -279,6 +279,7 @@ default_options_uncolored_help = (
     r"  -v, --verbose           Increase the default WARNING verbosity by one level\n"
     r"                          for each additional repetition of the option.\n"
     r"                          \[default: 0\]\n"
+    r"  --man                   Show the command's man page \(roff\) and exit.\n"
     r"  --version               Show the version and exit.\n"
     r"  -h, --help              Show this message and exit.\n"
 )
@@ -311,6 +312,7 @@ default_options_colored_help = (
     r"  \x1b\[36m\x1b\[1m-v\x1b\[0m, \x1b\[36m\x1b\[1m--verbose\x1b\[0m           Increase the default \x1b\[35m\x1b\[1mWARNING\x1b\[0m verbosity by one level\n"
     r"                          for each additional repetition of the option.\n"
     r"                          \x1b\[2m\[\x1b\[0m\x1b\[2mdefault: \x1b\[0m\x1b\[32m\x1b\[2m\x1b\[3m0\x1b\[0m\x1b\[2m\]\x1b\[0m\n"
+    r"  \x1b\[36m\x1b\[1m--man\x1b\[0m                   Show the command's man page \(roff\) and exit.\n"
     r"  \x1b\[36m\x1b\[1m--version\x1b\[0m               Show the version and exit.\n"
     r"  \x1b\[36m\x1b\[1m-h\x1b\[0m, \x1b\[36m\x1b\[1m--help\x1b\[0m              Show this message and exit.\n"
 )